test general properties for activations
