+++ tensorflow_addons/activations/gelu_test.py
@@ -64,19 +64,6 @@ class GeluTest(tf.test.TestCase, parameterized.TestCase):
             x = tf.ones(shape=shape, dtype=tf.float32)
             self.assertAllClose(fn(x), gelu(x))
 
-    def test_serialization(self):
-        config = tf.keras.activations.serialize(gelu)
-        fn = tf.keras.activations.deserialize(config)
-        self.assertEqual(fn, gelu)
-
-    def test_serialization_with_layers(self):
-        layer = tf.keras.layers.Dense(3, activation=gelu)
-        config = tf.keras.layers.serialize(layer)
-        deserialized_layer = tf.keras.layers.deserialize(config)
-        self.assertEqual(deserialized_layer.__class__.__name__,
-                         layer.__class__.__name__)
-        self.assertEqual(deserialized_layer.activation.__name__, "gelu")
-
 
 if __name__ == "__main__":
     tf.test.main()
